Perf: Temporary tile-draw reordering to better performance
That giant set of if statements in the level draw function slows down
performance, especially when combined with drawing a bunch of walls on
the screen.  There are better performance fixes, but for now, we've
shoved the problem to the side by making the if statements into an
if-elif chain and placed walls at the very top of that chain.