Feature: Make spawn, goal, and checkpoints place singletons in rect mode
Since there is only one spawn and one goal, it doesn't make sense to
place them in rect mode.  However, it's often a nuisance switching back
to place mode every time you move them, so I let them be placed as a
singleton even in rect mode.
Due to the way the checkpoint constructor works (requiring its column
and row to be passed as parameters), it didn't work with add_rect().
So, I made it place a singleton as well.